test(phone/settings): add unit tests for setting state (#551)
