+++ phone/package.json
@@ -15,10 +15,10 @@
     "dayjs": "^1.10.7",
     "dompurify": "^2.3.3",
     "emoji-mart": "^3.0.1",
-    "enzyme": "^3.11.0",
     "fivem-nui-react-lib": "^2.2.8",
     "howler": "^2.2.1",
     "i18next": "^20.6.1",
+    "jest": "^27.4.7",
     "jsonschema": "^1.4.0",
     "parse-url": "^5.0.7",
     "qs": "^6.9.6",
@@ -36,16 +36,17 @@
     "@sentry/react": "^5.29.2",
     "@sentry/tracing": "^5.29.2",
     "@sentry/webpack-plugin": "^1.18.3",
-    "@types/enzyme": "^3.10.10",
+    "@testing-library/jest-dom": "^5.16.1",
+    "@testing-library/react": "^12.1.2",
+    "@testing-library/react-hooks": "^7.0.2",
     "@types/howler": "^2.2.4",
-    "@types/jest": "^27.0.3",
+    "@types/jest": "^27.4.0",
     "@types/node": "^14.17.34",
     "@types/qs": "^6.9.5",
     "@types/react": "^16.14.21",
     "@types/react-dom": "^16.9.8",
     "@types/react-router-dom": "^5.3.2",
     "@welldone-software/why-did-you-render": "^6.2.3",
-    "@wojtekmaj/enzyme-adapter-react-17": "^0.6.5",
     "craco-alias": "^3.0.1",
     "cross-env": "^7.0.3",
     "dotenv": "^10.0.0",
@@ -62,6 +63,16 @@
   },
   "eslintConfig": {
     "extends": "react-app"
+  },
+  "jest": {
+    "moduleNameMapper": {
+      "^@apps(.*)$": "<rootDir>/src/apps$1",
+      "^@os(.*)$": "<rootDir>/src/os$1",
+      "^@ui(.*)$": "<rootDir>/src/ui$1",
+      "^@utils(.*)$": "<rootDir>/src/utils$1",
+      "^@common(.*)$": "<rootDir>/src/common$1",
+      "^@typings(.*)$": "<rootDir>/../typings$1"
+    }
   },
   "browserslist": {
     "production": [
