+++ phone/src/Phone.tsx
@@ -23,44 +23,27 @@ import { useTranslation } from 'react-i18next';
 import { PhoneEvents } from '@typings/phone';
 import { useKeyboardService } from '@os/keyboard/hooks/useKeyboardService';
 import PhoneWrapper from './PhoneWrapper';
-
 import dayjs from 'dayjs';
 import DefaultConfig from '../../config.json';
 import { TopLevelErrorComponent } from '@ui/components/TopLevelErrorComponent';
 import { useConfig } from '@os/phone/hooks/useConfig';
 import { useContactsListener } from './apps/contacts/hooks/useContactsListener';
 import { useNoteListener } from './apps/notes/hooks/useNoteListener';
-import { useSnackbar } from '@os/snackbar/hooks/useSnackbar';
 import { PhoneSnackbar } from '@os/snackbar/components/PhoneSnackbar';
-import { useCall } from './os/call/hooks/useCall';
-import { isSettingsSchemaValid } from './apps/settings/state/settings.state';
+import { useInvalidSettingsHandler } from './apps/settings/hooks/useInvalidSettingsHandler';
 
 function Phone() {
-  const { t, i18n } = useTranslation();
+  const { i18n } = useTranslation();
 
   const { apps } = useApps();
 
   const [settings] = useSettings();
-
-  const { addAlert } = useSnackbar();
 
   // Set language from local storage
   // This will only trigger on first mount & settings changes
   useEffect(() => {
     i18n.changeLanguage(settings.language.value).catch((e) => console.error(e));
   }, [i18n, settings.language]);
-
-  useEffect(() => {
-    if (!isSettingsSchemaValid()) {
-      addAlert({
-        message: t('SETTINGS.MESSAGES.INVALID_SETTINGS'),
-        type: 'error',
-      });
-    }
-    // We only want to run this on first mount of the phone,
-    // so we leave an empty dep array. Otherwise, we hit a max depth error.
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
 
   useConfig();
 
@@ -76,11 +59,9 @@ function Phone() {
   /*usePhotoService();*/
   useCallService();
   useDialService();
+  useInvalidSettingsHandler();
 
   const { modal: callModal } = useCallModal();
-  const { call } = useCall();
-
-  const showNavigation = call?.is_accepted || !callModal;
 
   return (
     <div>
@@ -99,7 +80,7 @@ function Phone() {
             <NotificationAlert />
             <PhoneSnackbar />
           </div>
-          {showNavigation && <Navigation />}
+          <Navigation />
         </PhoneWrapper>
       </TopLevelErrorComponent>
     </div>
