+++ phone/src/apps/settings/utils/constants.ts
@@ -1,0 +1,1 @@
+export const NPWD_STORAGE_KEY = 'npwd_settings';
