+++ phone/src/setupTests.ts
@@ -1,4 +1,2 @@
-import { configure } from 'enzyme';
-import Adapter from '@wojtekmaj/enzyme-adapter-react-17';
-
-configure({ adapter: new Adapter() });
+import '@testing-library/jest-dom';
+import './i18n';
